+++ src/emulator/functionsEmulatorRuntime.ts
@@ -495,14 +495,9 @@ async function initializeFirebaseAdminStubs(frb: FunctionsRuntimeBundle): Promis
       // If initializeApp() is called with options we use the provided options, otherwise
       // we use the default options.
       const defaultAppOptions = opts ? opts : defaultConfig;
-      new EmulatorLog(
-        "SYSTEM",
-        "default-admin-app-used",
-        `config=${JSON.stringify(defaultAppOptions)}`,
-        {
-          opts: defaultAppOptions,
-        }
-      ).log();
+      new EmulatorLog("SYSTEM", "default-admin-app-used", `config=${defaultAppOptions}`, {
+        opts: defaultAppOptions,
+      }).log();
 
       const defaultApp: admin.app.App = makeProxiedFirebaseApp(
         frb,
@@ -654,6 +649,32 @@ function initializeEnvironmentalVariables(frb: FunctionsRuntimeBundle): void {
     process.env.PUBSUB_EMULATOR_HOST = pubsubHost;
     logDebug(`Set PUBSUB_EMULATOR_HOST to ${pubsubHost}`);
   }
+}
+
+async function initializeFunctionsConfigHelper(frb: FunctionsRuntimeBundle): Promise<void> {
+  const functionsResolution = await requireResolveAsync("firebase-functions", {
+    paths: [frb.cwd],
+  });
+
+  const ff = require(functionsResolution);
+  logDebug("Checked functions.config()", {
+    config: ff.config(),
+  });
+
+  const originalConfig = ff.config();
+  const proxiedConfig = new Proxied(originalConfig)
+    .any((parentConfig, parentKey) => {
+      if (!parentConfig[parentKey]) {
+        new EmulatorLog("SYSTEM", "functions-config-missing-value", "", {
+          key: parentKey,
+        }).log();
+      }
+
+      return parentConfig[parentKey];
+    })
+    .finalize();
+
+  ff.config = () => proxiedConfig;
 }
 
 /**
@@ -932,6 +953,7 @@ async function initializeRuntime(
   }
 
   initializeNetworkFiltering(frb);
+  await initializeFunctionsConfigHelper(frb);
   await initializeFirebaseFunctionsStubs(frb);
   await initializeFirebaseAdminStubs(frb);