+++ src/test/emulators/functionsEmulatorRuntime.spec.ts
@@ -436,6 +436,42 @@ describe("FunctionsEmulator-Runtime", () => {
     });
   });
 
+  describe("_InitializeFunctionsConfigHelper()", () => {
+    it("should tell the user if they've accessed a non-existent function field", async () => {
+      const worker = InvokeRuntimeWithFunctions(
+        FunctionRuntimeBundles.onCreate,
+        () => {
+          require("firebase-admin").initializeApp();
+          return {
+            function_id: require("firebase-functions")
+              .firestore.document("test/test")
+              .onCreate(async () => {
+                /* tslint:disable:no-console */
+
+                // Exists
+                console.log(require("firebase-functions").config().real);
+
+                // Does not exist
+                console.log(require("firebase-functions").config().foo);
+                console.log(require("firebase-functions").config().bar);
+              }),
+          };
+        },
+        {
+          nodeBinary: process.execPath,
+          env: {
+            CLOUD_RUNTIME_CONFIG: JSON.stringify({
+              real: { exist: "already exists" },
+            }),
+          },
+        }
+      );
+
+      const logs = await _countLogEntries(worker);
+      expect(logs["functions-config-missing-value"]).to.eq(2);
+    }).timeout(TIMEOUT_MED);
+  });
+
   describe("Runtime", () => {
     describe("HTTPS", () => {
       it("should handle a GET request", async () => {